Refactor wkb linestring/linearring iterators in geo-generic-tests to have less code duplication
